loadtest/requirements.txt
locust>=2.17.0
psycopg2-binary>=2.9.9
redis>=5.0.0

